refactor: selectOptions 항목 utils로 이동
